VIL-924: adpat phases with same name as V1
